+++ Sources/TMDB_TVShowDetail/Entities/Creator.swift
@@ -1,0 +1,5 @@
+import Foundation
+struct Creator {
+    let name: String
+    let imageURL: String
+}
