+++ Sources/TMDB_TVShowDetail/Entities/Details.swift
@@ -1,0 +1,9 @@
+import Foundation
+struct Details {
+    let numberOfSeasons: Int
+    let numberOfEpisodes: Int
+    let firstAirDate: String
+    let lastAirDate: String
+    let status: String
+    let averageVote: String
+}
