+++ Sources/TMDB_TVShowDetail/Entities/Season.swift
@@ -1,0 +1,8 @@
+import Foundation
+struct Season: Identifiable {
+    let id: String
+    let name: String
+    let posterURL: String
+    let episodeCount: Int
+    let airDate: String
+}
