+++ Sources/TMDB_TVShowDetail/Entities/ShowData.swift
@@ -1,0 +1,11 @@
+import Foundation
+struct ShowData {
+    let title: String
+    let tagline: String
+    let posterURL: String
+    let overview: String
+    let genres: String
+    let creators: [Creator]
+    let details: Details
+    let seasons: [Season]
+}
